# backend/ecollecto-backend/openapi.yaml
openapi: 3.1.0
info:
  title: eCollecto API
  description: eCollecto philatelic collection REST API. Public catalog endpoints
    are accessible without authentication. Protected endpoints require a valid Bearer
    JWT issued by Keycloak.
  version: v1
servers:
- url: http://localhost:8080
  description: Local development server
...
paths:
...
  /api/me/favorites/items:
    post:
      tags:
      - Favorites
      summary: Add to favorites
      description: Adds a stamp to the authenticated user's favorites.
      operationId: addToFavorites
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteItemRequest"
        required: true
      responses:
        "201":
          description: Stamp added to favorites.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FavoriteItemDto"
        "400":
          description: Invalid request body.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Stamp already in favorites.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
  /api/me/collection/items:
    post:
      tags:
      - Collection
      summary: Add to collection
      description: Adds a stamp to the authenticated user's collection.
      operationId: addToCollection
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCollectionItemRequest"
        required: true
      responses:
        "201":
          description: Stamp added to collection.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollectionItemDto"
        "400":
          description: Invalid request body.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Stamp already in collection.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
...
  /api/me/favorites:
    get:
      tags:
      - Favorites
      summary: Get favorites
      description: Returns all stamps in the authenticated user's favorites.
      operationId: getFavorites
      responses:
        "200":
          description: Favorites retrieved.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoriteItemDto"
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
  /api/me/collection:
    get:
      tags:
      - Collection
      summary: Get collection
      description: Returns all stamps in the authenticated user's collection.
      operationId: getCollection
      responses:
        "200":
          description: Collection retrieved.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionItemDto"
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
...
  /api/me/favorites/items/{stampId}:
    delete:
      tags:
      - Favorites
      summary: Remove from favorites
      description: Removes a stamp from the authenticated user's favorites.
      operationId: removeFromFavorites
      parameters:
      - name: stampId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Stamp removed from favorites.
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Stamp not in favorites.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
  /api/me/collection/items/{stampId}:
    delete:
      tags:
      - Collection
      summary: Remove from collection
      description: Removes a stamp from the authenticated user's collection.
      operationId: removeFromCollection
      parameters:
      - name: stampId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Stamp removed from collection.
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Stamp not in collection.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
components:
  schemas:
    ErrorResponse:
      type: object
      description: Error response payload.
      properties:
        message:
          type: string
          description: Error message.
        code:
          type: string
          description: Error code.
        status:
          type: integer
          format: int32
          description: HTTP status code.
...
    AddFavoriteItemRequest:
      type: object
      description: Request body to add a stamp to the user's favorites.
      properties:
        stampId:
          type: string
          description: Stamp ID to add.
          example: stamp-2024-001
          minLength: 1
      required:
      - stampId
    FavoriteItemDto:
      type: object
      description: A stamp in the user's favorites.
      properties:
        stampId:
          type: string
          description: Stamp ID (references stamps._id).
          example: stamp-2024-001
        addedAt:
          type: string
          format: date-time
          description: Timestamp when the stamp was added to favorites.
    AddCollectionItemRequest:
      type: object
      description: Request body to add a stamp to the user's collection.
      properties:
        stampId:
          type: string
          description: Stamp ID to add.
          example: stamp-2024-001
          minLength: 1
      required:
      - stampId
    CollectionItemDto:
      type: object
      description: A stamp in the user's collection.
      properties:
        stampId:
          type: string
          description: Stamp ID (references stamps._id).
          example: stamp-2024-001
        addedAt:
          type: string
          format: date-time
          description: Timestamp when the stamp was added to the collection.
...
  securitySchemes:
    openIdConnect:
      type: openIdConnect
      description: OpenID Connect discovery for the eCollecto Keycloak realm. Use
        the ecollecto-ui public client with PKCE (S256) for the browser flow.
      openIdConnectUrl: http://localhost:8180/realms/ecollecto/.well-known/openid-configuration
    bearerAuth:
      type: http
      description: "JWT Bearer token issued by Keycloak (Authorization Code + PKCE).\
        \ Obtain a token from: http://localhost:8180/realms/ecollecto/protocol/openid-connect/token"
      scheme: bearer
      bearerFormat: JWT